+++ source/rr/src/gameexec.cpp
@@ -860,7 +860,7 @@ static void VM_AddInventory(DukePlayer_t * const pPlayer, int const itemNum, int
     case GET_SHIELD:
     {
         int16_t & shield_amount = pPlayer->inv_amount[GET_SHIELD];
-        shield_amount = min(shield_amount + nAmount, pPlayer->max_shield_amount);
+        shield_amount = min(shield_amount + nAmount, (int)pPlayer->max_shield_amount);
         break;
     }
